refactor some more
